queuing in import
