+++ app/Jobs/NotifyUserOfCompletedImport.php
@@ -1,0 +1,37 @@
+<?php
+
+namespace App\Jobs;
+
+use App\Notifications\MemeberImportExcelNotification;
+use App\Person;
+use Illuminate\Bus\Queueable;
+use Illuminate\Contracts\Queue\ShouldQueue;
+use Illuminate\Foundation\Bus\Dispatchable;
+use Illuminate\Queue\InteractsWithQueue;
+use Illuminate\Queue\SerializesModels;
+
+class NotifyUserOfCompletedImport implements ShouldQueue
+{
+    use Dispatchable, InteractsWithQueue, Queueable, SerializesModels;
+
+    public $person;
+    /**
+     * Create a new job instance.
+     *
+     * @return void
+     */
+    public function __construct(Person $person)
+    {
+        $this->person = $person;
+    }
+
+    /**
+     * Execute the job.
+     *
+     * @return void
+     */
+    public function handle()
+    {
+        $this->person->notify(new MemeberImportExcelNotification());
+    }
+}
