+++ resources/lang/es/messages.php
@@ -751,6 +751,7 @@ return [
         'please_wait'                  => 'es-Please wait while this content loads.',
         'org_default_update_success'   => 'es-Your default Organization Updated',
         'new_org_created_successfully' => 'es-New Org Created Successfully.',
+        'import_file_queued'           => 'es-Import is queued. You will be notified via email when its completed.',
     ],
 
     'modals'                 => [
@@ -886,6 +887,12 @@ return [
             'line1'   => 'es-An mCentric account was setup for you by :name with :org.',
             'line2'   => 'es-Your login is the email address where you are receiving this notification.',
             'line3'   => 'es-The password that was set for you is: \':pass\'',
+        ],
+
+        //Member import complete
+        'member_import' => [
+            'subject' => 'es-Import completed.',
+            'line1'   => 'es-Your member import is completed.',
         ],
 
     ],
